move function to common, rather than int values, #219
