+++ tests/unit/vis/_vis_utils/_ms/_tables/test_read.py
@@ -291,23 +291,20 @@ def test_get_pad_value_uvw(main_xds_min):
 
 
 def test_get_pad_value_feed1(main_xds_min):
-    from xradio.vis._vis_utils._ms._tables.read import get_pad_value
-    from xradio._utils.common import fill_value_int32
+    from xradio._utils.common import get_pad_value
 
     res = get_pad_value(main_xds_min.data_vars["feed1_id"].dtype)
 
-    assert res == fill_value_int32
+    assert res == get_pad_value(np.int32)
 
 
 def test_get_pad_value_state_id(main_xds_min):
-    from xradio.vis._vis_utils._ms._tables.read import get_pad_value
-    from xradio._utils.common import fill_value_int64
+    from xradio._utils.common import get_pad_value
 
-    print(f"{main_xds_min.data_vars=}")
     # In the xds from xds_helper all ints are turned into int32, still check int64
     res = get_pad_value(main_xds_min.data_vars["STATE_ID"].astype(np.int64).dtype)
 
-    assert res == fill_value_int64
+    assert res == get_pad_value(np.int64)
 
 
 def test_redimension_ms_subtable_source(source_xds_min):
